feat: implement user and organization management middleware with Stripe integration and default chatflow creation
