+++ packages/server/src/middlewares/authentication/index.ts
@@ -1,12 +1,13 @@
 import { NextFunction, Request, Response } from 'express'
 import { auth } from 'express-oauth2-jwt-bearer'
-import Stripe from 'stripe'
 
 import { DataSource } from 'typeorm'
 import { User } from '../../database/entities/User'
-import { Organization } from '../../database/entities/Organization'
 import apikeyService from '../../services/apikey'
-import { QueryFailedError } from 'typeorm'
+import { findOrCreateOrganization } from './findOrCreateOrganization'
+import { findOrCreateUser } from './findOrCreateUser'
+import { ensureStripeCustomerForUser } from './ensureStripeCustomerForUser'
+import { findOrCreateDefaultChatflowsForUser } from './findOrCreateDefaultChatflowsForUser'
 
 const jwtCheck = auth({
     authRequired: true,
@@ -54,144 +55,10 @@ const tryApiKeyAuth = async (req: Request, AppDataSource: DataSource): Promise<U
     }
 }
 
-/**
- * Safely creates or updates a user with proper transaction handling to prevent race conditions
- */
-const findOrCreateUser = async (
-    AppDataSource: DataSource,
-    auth0Id: string,
-    email: string,
-    name: string,
-    organizationId: string
-): Promise<User> => {
-    const userRepo = AppDataSource.getRepository(User)
-
-    // First, try to find the user
-    let user = await userRepo.findOneBy({ auth0Id })
-    if (user) {
-        // Update user if needed and return
-        if (user.email !== email || user.name !== name || user.organizationId !== organizationId) {
-            user.email = email
-            user.name = name
-            user.organizationId = organizationId
-            await userRepo.save(user)
-        }
-        return user
-    }
-
-    // User not found, try to create with transaction and retry logic
-    let retryCount = 0
-    const maxRetries = 3
-
-    while (retryCount < maxRetries) {
-        const queryRunner = AppDataSource.createQueryRunner()
-        await queryRunner.connect()
-        await queryRunner.startTransaction()
-
-        try {
-            // Check again inside transaction to prevent race condition
-            const existingUser = await queryRunner.manager.findOneBy(User, { auth0Id })
-            if (existingUser) {
-                await queryRunner.release()
-                return existingUser
-            }
-
-            // Create new user
-            const newUser = userRepo.create({ auth0Id, email, name, organizationId })
-            const savedUser = await queryRunner.manager.save(newUser)
-
-            await queryRunner.commitTransaction()
-            await queryRunner.release()
-
-            console.log(`[AuthMiddleware] Successfully created new user: ${savedUser.id}`)
-            return savedUser
-        } catch (error) {
-            await queryRunner.rollbackTransaction()
-            await queryRunner.release()
-
-            // If duplicate key error, retry after a short delay
-            if (error instanceof QueryFailedError && error.message.includes('duplicate key')) {
-                retryCount++
-                console.log(`[AuthMiddleware] Duplicate key error, retrying (${retryCount}/${maxRetries})`)
-                await new Promise((resolve) => setTimeout(resolve, 100 * retryCount)) // Exponential backoff
-
-                // Try to get the user that was created in parallel
-                user = await userRepo.findOneBy({ auth0Id })
-                if (user) {
-                    return user
-                }
-            } else {
-                throw error
-            }
-        }
-    }
-
-    // Last attempt to find user after retries
-    user = await userRepo.findOneBy({ auth0Id })
-    if (user) {
-        return user
-    }
-
-    throw new Error(`Failed to create or find user with auth0Id: ${auth0Id} after ${maxRetries} retries`)
-}
-
-/**
- * Safely creates or updates an organization with proper transaction handling
- */
-const findOrCreateOrganization = async (AppDataSource: DataSource, auth0OrgId: string, orgName: string): Promise<Organization> => {
-    const orgRepo = AppDataSource.getRepository(Organization)
-
-    let organization = await orgRepo.findOneBy({ auth0Id: auth0OrgId })
-
-    if (organization) {
-        if (organization.name !== orgName) {
-            organization.name = orgName
-            await orgRepo.save(organization)
-        }
-        return organization
-    }
-
-    // Organization not found, create with transaction
-    const queryRunner = AppDataSource.createQueryRunner()
-    await queryRunner.connect()
-    await queryRunner.startTransaction()
-
-    try {
-        // Check again inside transaction
-        organization = await queryRunner.manager.findOneBy(Organization, { auth0Id: auth0OrgId })
-        if (organization) {
-            await queryRunner.release()
-            return organization
-        }
-
-        // Create new organization
-        const newOrg = orgRepo.create({ auth0Id: auth0OrgId, name: orgName })
-        organization = await queryRunner.manager.save(newOrg)
-
-        await queryRunner.commitTransaction()
-        await queryRunner.release()
-
-        return organization
-    } catch (error) {
-        await queryRunner.rollbackTransaction()
-        await queryRunner.release()
-
-        if (error instanceof QueryFailedError && error.message.includes('duplicate key')) {
-            // If duplicate key, try to fetch the existing organization
-            organization = await orgRepo.findOneBy({ auth0Id: auth0OrgId })
-            if (organization) {
-                return organization
-            }
-        }
-
-        throw error
-    }
-}
-
 export const authenticationHandlerMiddleware =
     ({ whitelistURLs, AppDataSource }: { whitelistURLs: string[]; AppDataSource: DataSource }) =>
     async (req: Request, res: Response, next: NextFunction) => {
-        const startTime = new Date().getTime()
+        // const startTime = new Date().getTime()
         const requireAuth = /\/api\/v1\//i.test(req.url) && !whitelistURLs.some((url) => req.url.includes(url))
         const jwtMiddleware = requireAuth ? jwtCheck : jwtCheckPublic
 
@@ -261,52 +128,14 @@ export const authenticationHandlerMiddleware =
                     const organization = await findOrCreateOrganization(AppDataSource, userOrgId, authUser.org_name)
 
                     // Get or create user using transaction-safe method
-                    const user = await findOrCreateUser(AppDataSource, auth0Id, email, name, organization.id)
+                    let user = await findOrCreateUser(AppDataSource, auth0Id, email, name, organization.id)
 
-                    // Upsert customer on Stripe if no customerId is attached
-                    let stripeCustomerId = user.stripeCustomerId
-                    if (!stripeCustomerId && process.env.BILLING_STRIPE_SECRET_KEY) {
-                        if (
-                            process.env.BILLING_STRIPE_ORGANIZATION_CUSTOMER_ID ||
-                            (organization.stripeCustomerId && organization.billingPoolEnabled == true)
-                        ) {
-                            stripeCustomerId = process.env.BILLING_STRIPE_ORGANIZATION_CUSTOMER_ID ?? organization.stripeCustomerId!
-                        } else {
-                            try {
-                                const stripe = new Stripe(process.env.BILLING_STRIPE_SECRET_KEY ?? '')
+                    // Replace the Stripe customer logic with the new ensureStripeCustomerForUser function
+                    user = await ensureStripeCustomerForUser(AppDataSource, user, organization, auth0Id, email, name)
 
-                                const existingCustomer = await stripe.customers.list({
-                                    email,
-                                    limit: 1
-                                })
+                    // Find or create default chatflows for the user
+                    await findOrCreateDefaultChatflowsForUser(AppDataSource, user)
 
-                                if (existingCustomer.data.length > 0) {
-                                    stripeCustomerId = existingCustomer.data[0].id
-                                } else {
-                                    const customer = await stripe.customers.create(
-                                        {
-                                            email,
-                                            name,
-                                            metadata: {
-                                                userId: user.id,
-                                                auth0Id,
-                                                orgId: organization.id
-                                            }
-                                        },
-                                        { idempotencyKey: auth0Id }
-                                    )
-                                    stripeCustomerId = customer.id
-                                }
-                                // Optionally, update the user profile in your database with the new customerId
-                                user.stripeCustomerId = stripeCustomerId
-                            } catch (error) {
-                                console.error('Error creating/updating Stripe customers:', error)
-                                return res.status(500).send('Internal Server Error')
-                            }
-                            await AppDataSource.getRepository(User).save(user)
-                        }
-                        user.stripeCustomerId = stripeCustomerId
-                    }
                     req.user = { ...authUser, ...user, roles }
                 } catch (error) {
                     console.error('[AuthMiddleware] Error during user authentication:', error)
